travis: run copr build only for the master branch
This is needed to avoid building into the @codescan/csdiff copr project
upon a push to a non-master branch.  Also, pull-requests against
a non-master branch would be excluded, but they should be rare.

Drop the -x flag from the Travis build.  That PR webhook isn't private
information but still..  It is better to not make a honeypot from all
our Travis logs.